Tests utils clean-up
