+++ dataland-frontend/tests/e2e/specs/eu-taxonomy-financials/AddRemoveKpis.ts
@@ -3,7 +3,7 @@ import { describeIf } from "@e2e/support/TestUtility";
 import { generateDummyCompanyInformation } from "@e2e/utils/CompanyUpload";
 import { admin_name, admin_pw } from "@e2e/utils/Cypress";
 import { fillAndValidateEuTaxonomyCreditInstitutionForm } from "@e2e/utils/EuTaxonomyFinancialsUpload";
-import { uploadCompanyViaApiAndEuTaxonomyDataViaForm } from "@e2e/utils/GeneralApiUtils";
+import { uploadCompanyViaApiAndEuTaxonomyDataViaForm } from "@e2e/utils/GeneralUtils";
 import { FixtureData, getPreparedFixture } from "@sharedUtils/Fixtures";
 
 describeIf(
