+++ dataland-frontend/tests/e2e/specs/lksg/index.ts
@@ -3,4 +3,5 @@
  */
 describe("Tests for view page of Lksg-Framework data", () => {
   require("./DataIntegrity");
+  require("./EditButtonWithMultipleReportingPeriods");
 });
