+++ dataland-frontend/tests/e2e/utils/GeneralApiUtils.ts
@@ -1,15 +1,5 @@
-import {
-  CompanyDataControllerApi,
-  CompanyInformation,
-  Configuration,
-  DataTypeEnum,
-  StoredCompany,
-} from "@clients/backend";
-import { CyHttpMessages, Interception, RouteHandler } from "cypress/types/net-stubbing";
-import Chainable = Cypress.Chainable;
-import { getKeycloakToken } from "./Auth";
-import { admin_name, admin_pw, uploader_name, uploader_pw } from "@e2e/utils/Cypress";
-import { generateDummyCompanyInformation, uploadCompanyViaApi } from "./CompanyUpload";
+import { CompanyDataControllerApi, Configuration, DataTypeEnum, StoredCompany } from "@clients/backend";
+import { RouteHandler } from "cypress/types/net-stubbing";
 
 export interface UploadIds {
   companyId: string;
@@ -97,78 +87,4 @@ export function interceptAllDataPostsAndBypassQaIfPossible(): void {
     }
   };
   cy.intercept("/api/data/*", handler);
-}
-
-/**
- * Visits the edit page for a framework via UI navigation.
- * @param companyId the id of the company for which to edit a dataset
- * @param dataType the framework type
- * @returns a cypress chainable to the interception of the data request on the edit page
- */
-export function goToEditFormOfMostRecentDataset(companyId: string, dataType: DataTypeEnum): Chainable<Interception> {
-  const getRequestAlias = "getData";
-  cy.intercept({
-    method: "GET",
-    url: "**/api/data/**",
-    times: 2,
-  }).as(getRequestAlias);
-  cy.visit(`/companies/${companyId}/frameworks/${dataType}`);
-  cy.wait(`@${getRequestAlias}`, { timeout: Cypress.env("medium_timeout_in_ms") as number });
-  cy.get('[data-test="editDatasetButton"]').click();
-  return cy.wait(`@${getRequestAlias}`, { timeout: Cypress.env("medium_timeout_in_ms") as number });
-}
-
-/**
- * Uploads a company via POST-request, then an EU Taxonomy dataset for financial companies for the uploaded company
- * via the form in the frontend, and then visits the view page where that dataset is displayed
- * @param frameworkDataType The EU Taxanomy framework being tested
- * @param companyInformation Company information to be used for the company upload
- * @param testData EU Taxonomy dataset for financial companies to be uploaded
- * @param formFill Steps involved to fill data of the upload form
- * @param submissionDataIntercept performs checks on the request itself
- * @param afterDatasetSubmission is performed after the data has been submitted
- */
-export function uploadCompanyViaApiAndEuTaxonomyDataViaForm<T>(
-  frameworkDataType: DataTypeEnum,
-  companyInformation: CompanyInformation,
-  testData: T,
-  formFill: (data: T) => void,
-  submissionDataIntercept: (request: CyHttpMessages.IncomingHttpRequest) => void,
-  afterDatasetSubmission: (companyId: string) => void
-): void {
-  getKeycloakToken(uploader_name, uploader_pw).then((token: string) => {
-    return uploadCompanyViaApi(token, generateDummyCompanyInformation(companyInformation.companyName)).then(
-      (storedCompany): void => {
-        cy.ensureLoggedIn(admin_name, admin_pw);
-        cy.visitAndCheckAppMount(`/companies/${storedCompany.companyId}/frameworks/${frameworkDataType}/upload`);
-        formFill(testData);
-        submitFilledInEuTaxonomyForm(submissionDataIntercept);
-        afterDatasetSubmission(storedCompany.companyId);
-      }
-    );
-  });
-}
-
-/**
- * After a Eu Taxonomy financial or non financial form has been filled in this function submits the form and checks
- * if a 200 response is returned by the backend
- * @param submissionDataIntercept function that asserts content of an intercepted request
- */
-export function submitFilledInEuTaxonomyForm(
-  submissionDataIntercept: (request: CyHttpMessages.IncomingHttpRequest) => void
-): void {
-  const postRequestAlias = "postDataAlias";
-  cy.intercept(
-    {
-      method: "POST",
-      url: `**/api/data/**`,
-      times: 1,
-    },
-    submissionDataIntercept
-  ).as(postRequestAlias);
-  cy.get('button[data-test="submitButton"]').click();
-  cy.wait(`@${postRequestAlias}`, { timeout: Cypress.env("long_timeout_in_ms") as number }).then((interception) => {
-    expect(interception.response?.statusCode).to.eq(200);
-  });
-  cy.contains("td", "EU Taxonomy");
 }
